hare3: fix overlow in multiplication
we are running with smaller limit (4 iteration) that fits uint8,
but we plan to eventually raise to 40

+++ hare3/types.go
@@ -61,7 +61,7 @@ func (ir IterRound) IsMessageRound() bool {
 }
 
 func (ir IterRound) Absolute() uint32 {
-	return uint32(ir.Iter*uint8(notify) + uint8(ir.Round))
+	return uint32(ir.Iter)*uint32(notify) + uint32(ir.Round)
 }
 
 type Value struct {
